requirements.txt
Django
celery
redis
gunicorn
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-celery-results
django-celery-beat